requirements.txt
asgiref==3.11.1
Django==4.2.10
django-cors-headers==4.3.1
django-environ==0.13.0
django-filter==23.5
django-guardian==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Markdown==3.10.2
psycopg2-binary==2.9.11
PyJWT==2.11.0
pytz==2025.2
sqlparse==0.5.5
